Update CKEditor4 html5 demo for a proof of concept
We've received a support request to validate whether or not, the MahtType
plugin for CKEditor4 is compatible with the Enhanced Image plugin.

The optional Enhanced Image plugin, introduced in CKEditor 4.3,
supports inserting images into the editor content.

The purpose of this research is to build a CKEditor4 demo using Enhance Image
and MathType plugins together in a way that solves and/or mitigates the issues
detected so far by the client and our own tests.

The premise is to use the CKEditor API while not making any change on any of
the plugins source code to integrate them togethere.

At the same time, any unexpected behavior on the MathType plugin resulting of
this experiment, will be registered as an issue in our backlog in order to be
fixed whenever possible.

The deliveries included on this commit are:

* An example demo and its source code.
* A step by step guide to integrate both plugins.

See: https://ckeditor.com/docs/ckeditor4/latest/examples/image2.html